created InlineGlobalHandler to handle inline queries
