+++ src/main/java/com/fbc/bot/telegram/handler/input/MessageType.java
@@ -4,8 +4,7 @@ public enum MessageType {
     START,
     SHARE_COCK_SIZE,
     JOIN_PARTY,
-    INLINE_SHARE_COCK_SIZE,
-    INLINE_UNKNOWN,
+    INLINE_GLOBAL,
     ADD_MUSIC,
     UNKNOWN
 }
