+++ src/main/java/com/fbc/bot/telegram/service/CockSizeService.java
@@ -1,9 +1,9 @@
 package com.fbc.bot.telegram.service;
 
 import com.fbc.bot.cocksize.model.DailyCockSize;
+import com.fbc.bot.cocksize.service.CockSizeGenerator;
 import com.fbc.bot.user.model.User;
 import com.fbc.bot.user.service.UserService;
-import com.fbc.bot.cocksize.service.CockSizeGenerator;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
