+++ src/main/java/com/fbc/bot/user/mapper/UserMapper.java
@@ -1,7 +1,7 @@
 package com.fbc.bot.user.mapper;
 
-import com.fbc.bot.user.model.User;
 import com.fbc.bot.user.dto.UserDto;
+import com.fbc.bot.user.model.User;
 import org.mapstruct.Mapper;
 import org.mapstruct.Mapping;
