+++ src/test/java/com/fbc/bot/common/CockSizeServiceTest.java
@@ -1,9 +1,9 @@
 package com.fbc.bot.common;
 
-import com.fbc.bot.user.service.UserService;
+import com.fbc.bot.telegram.service.CockSizeService;
 import com.fbc.bot.user.dto.UserDto;
 import com.fbc.bot.user.model.User;
-import com.fbc.bot.telegram.service.CockSizeService;
+import com.fbc.bot.user.service.UserService;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
